An app that reads a txt with numbers and produces all possible and unique 6 number combinations
The path of the file should be update to your own
Reads with a Scanner and a while loop to iterate through all number
till the -1 break point

Uses boolean static methods as hasLessThan4EvensOrOdds
to filter out some combinations

sixIntCombination is the method that produce the 6 number combination
with the use of 6 for nested loops. After a combination is produced
it is filtered and if it passes the it is printed into a txt file

The printArrayToFile method uses a bufferWriter wrapped around a FileWriter
to print the array to a txt file. Append is true so it can print
all the possible combinations.